Add records of session establishment for subscription resumption

+++ src/app/SubscriptionResumptionSessionEstablisher.cpp
@@ -36,6 +36,31 @@ private:
     SubscriptionResumptionSessionEstablisher * mSessionEstablisher;
 };
 
+struct SubscriptionResumptionSessionEstablishRecord {
+    chip::SubscriptionId mSubscriptionId;
+    uint16_t mSessionEstablishRetries = UINT16_MAX;
+};
+
+static SubscriptionResumptionSessionEstablishRecord sSubscriptionResumptionRecords[CHIP_IM_MAX_NUM_SUBSCRIPTIONS];
+
+static SubscriptionResumptionSessionEstablishRecord *FindRecordOrFirstEmptyEntry(chip::SubscriptionId subscriptionId)
+{
+    SubscriptionResumptionSessionEstablishRecord *ret = nullptr;
+    for (size_t i = 0; i < CHIP_IM_MAX_NUM_SUBSCRIPTIONS; ++i)
+    {
+        if (sSubscriptionResumptionRecords[i].mSubscriptionId == subscriptionId)
+        {
+            ret = &sSubscriptionResumptionRecords[i];
+            break;
+        }
+        if (sSubscriptionResumptionRecords[i].mSessionEstablishRetries == UINT16_MAX && !ret)
+        {
+            ret = &sSubscriptionResumptionRecords[i];
+        }
+    }
+    return ret;
+}
+
 SubscriptionResumptionSessionEstablisher::SubscriptionResumptionSessionEstablisher() :
     mOnConnectedCallback(HandleDeviceConnected, this), mOnConnectionFailureCallback(HandleDeviceConnectionFailure, this)
 {}
@@ -86,6 +111,11 @@ void SubscriptionResumptionSessionEstablisher::HandleDeviceConnected(void * cont
 {
     AutoDeleteEstablisher establisher(static_cast<SubscriptionResumptionSessionEstablisher *>(context));
     SubscriptionResumptionStorage::SubscriptionInfo & subscriptionInfo = establisher->mSubscriptionInfo;
+    SubscriptionResumptionSessionEstablishRecord *record = FindRecordOrFirstEmptyEntry(subscriptionInfo.mSubscriptionId) ;
+    if (record && record->mSubscriptionId == subscriptionInfo.mSubscriptionId)
+    {
+        record->mSessionEstablishRetries = UINT16_MAX;
+    }
     InteractionModelEngine * imEngine                                  = InteractionModelEngine::GetInstance();
     if (!imEngine->EnsureResourceForSubscription(subscriptionInfo.mFabricIndex, subscriptionInfo.mAttributePaths.AllocatedSize(),
                                                  subscriptionInfo.mEventPaths.AllocatedSize()))
@@ -109,14 +139,21 @@ void SubscriptionResumptionSessionEstablisher::HandleDeviceConnectionFailure(voi
     SubscriptionResumptionStorage::SubscriptionInfo & subscriptionInfo = establisher->mSubscriptionInfo;
     ChipLogError(DataManagement, "Failed to establish CASE for subscription-resumption with error '%" CHIP_ERROR_FORMAT "'",
                  error.Format());
-    // If the device fails to establish the session, the subscriber might be offline and its subscription read client will
-    // be deleted when the device reconnect to the subscriber. This subscription will be never used again. So clean up
-    // the persistent subscription information storage.
-    auto * subscriptionResumptionStorage = InteractionModelEngine::GetInstance()->GetSubscriptionResumptionStorage();
-    if (subscriptionResumptionStorage)
+    SubscriptionResumptionSessionEstablishRecord *record = FindRecordOrFirstEmptyEntry(subscriptionInfo.mSubscriptionId) ;
+    VerifyOrReturn(record);
+    record->mSubscriptionId = subscriptionInfo.mSubscriptionId;
+    record->mSessionEstablishRetries = record->mSessionEstablishRetries == UINT16_MAX ? 1 : record->mSessionEstablishRetries + 1;
+    if (record->mSessionEstablishRetries > CHIP_CONFIG_SUBSCRIPTION_TIMEOUT_RESUMPTION_MAX_FIBONACCI_STEP_INDEX)
     {
-        subscriptionResumptionStorage->Delete(subscriptionInfo.mNodeId, subscriptionInfo.mFabricIndex,
-                                              subscriptionInfo.mSubscriptionId);
+        // If the device fails to establish the session for several times, the subscriber might be offline and its subscription
+        // read client will be deleted when the device reconnect to the subscriber. This subscription will be never used again.
+        // So clean up the persistent subscription information storage.
+        auto * subscriptionResumptionStorage = InteractionModelEngine::GetInstance()->GetSubscriptionResumptionStorage();
+        if (subscriptionResumptionStorage)
+        {
+            subscriptionResumptionStorage->Delete(subscriptionInfo.mNodeId, subscriptionInfo.mFabricIndex,
+                                                  subscriptionInfo.mSubscriptionId);
+        }
     }
 }
